requirements.txt
python-telegram-bot==20.0
wakeonlan==3.0.0
